RAGAS/requirements.txt
# Основные зависимости для RAG проекта
langchain==0.1.0
langchain-community==0.0.10
langchain-huggingface==0.0.1

# ML/DL библиотеки
torch>=2.0.0
transformers>=4.35.0
sentence-transformers>=2.2.2
accelerate>=0.24.0

# Векторные базы данных
chromadb>=0.4.18
faiss-cpu>=1.7.4

# Трекинг экспериментов
mlflow>=2.8.1

# Веб-интерфейс
gradio>=4.0.0
streamlit>=1.28.0

# Работа с данными
datasets>=2.14.0
pandas>=2.0.0
numpy>=1.24.0

# Конфигурация
pyyaml>=6.0.1

# Квантизация (опционально) - установите только при необходимости
# llama-cpp-python>=0.2.0  # Раскомментируйте если нужна поддержка GGUF
# auto-gptq>=0.4.0  # Раскомментируйте если нужна GPTQ квантизация (требует CUDA)
# bitsandbytes>=0.41.0  # Раскомментируйте если нужна 8-bit квантизация (требует CUDA)

# Дополнительные утилиты
tqdm>=4.65.0
scikit-learn>=1.3.0
rouge-score>=0.1.2
sacrebleu>=2.3.1
nltk>=3.8.1

# Логирование и мониторинг
wandb>=0.16.0
tensorboard>=2.14.0

# Документооборот
pypdf>=3.17.0
python-docx>=0.8.11
python-multipart>=0.0.6

# HTTP клиент
httpx>=0.25.0
requests>=2.31.0
